Battle System Update:
Started the starting stage of death state for character. Implemented UI element to display an enemy's charge diamond count. Each turn, the charge diamond bar fills up. Once all is full, the enemy will perform a special attack on the next turn. Once they have performed this special attack, all charge diamonds are reset.

+++ Client/Assets/Scripts/System/Battle/EnemyBase.cs
@@ -14,7 +14,7 @@ public class EnemyBase
     public float currentAtk;
     public float baseDef;
     public float currentDef;
-    public float ChargeDiamond;
+    public int ChargeDiamond;
     public enum EleType { FIRE, WATER, EARTH, WIND, DARK, LIGHT }
     public EleType EnemyEle;
     public List<BaseSkill> mySkill;
